Sets up model and controller for Mangas
Previously there was no controller or model for the manga portion
of this app. This commit instantiates the Manga model as well as
the controller. The controller currently only has two actions:
`index`, for showing all mangas, and `show`, for display a particular
manga instance.

Furthermore, this commit exposes the index and show routes so that
information may be fetched via GET request.